+++ application/comit_node/Cargo.toml
@@ -35,6 +35,7 @@ logging = { path = "../logging" }
 event_store = { path = "../../vendor/event_store" }
 rand = "0.4.2"
 rustc-hex = "1.0.0"
+ganp = { path = "../ganp" }
 
 [dev-dependencies]
 env_logger = "0.5.11"
