+++ application/comit_node/src/ledger_htlc_service.rs
@@ -1,7 +1,5 @@
-use common_types::{
-    ledger::Ledger,
-    secret::{Secret, SecretHash},
-};
+use common_types::secret::{Secret, SecretHash};
+use ganp::ledger::Ledger;
 use secp256k1_support::KeyPair;
 use swaps::common::TradeId;
