+++ application/comit_node/src/swaps/eth_btc/cli/buy_routes.rs
@@ -1,13 +1,10 @@
 use bitcoin_htlc::{self, Htlc as BtcHtlc};
 use bitcoin_support::{self, BitcoinQuantity, Blocks, Network, PubkeyHash};
 use comit_node_api_client::{ApiClient, OrderRequestBody};
-use common_types::{
-    ledger::{bitcoin::Bitcoin, ethereum::Ethereum, Ledger},
-    secret::Secret,
-    TradingSymbol,
-};
+use common_types::{secret::Secret, TradingSymbol};
 use ethereum_support::{self, EthereumQuantity};
 use event_store::{EventStore, InMemoryEventStore};
+use ganp::ledger::{bitcoin::Bitcoin, ethereum::Ethereum, Ledger};
 use rand::OsRng;
 use rocket::{response::status::BadRequest, State};
 use rocket_contrib::Json;
