+++ application/ganp/Cargo.toml
@@ -16,6 +16,7 @@ common_types = { path = "../common_types" }
 log = "0.4"
 serde = "1"
 serde_derive = "1"
+bitcoin_rpc_client = "0.1"
 
 [dev-dependencies]
 memsocket = "0.1"
