+++ application/ganp/src/lib.rs
@@ -6,8 +6,10 @@ extern crate common_types;
 extern crate log;
 #[macro_use]
 extern crate serde_derive;
+extern crate bitcoin_rpc_client;
 extern crate bitcoin_support;
 extern crate ethereum_support;
+extern crate secp256k1_support;
 extern crate serde;
 
 mod config;
@@ -15,5 +17,6 @@ mod handler;
 pub mod rfc003;
 pub mod swap;
 
+pub mod ledger;
 pub use config::*;
 pub use handler::*;
